🔧 Fix custom schemes.
